requirements.txt
pytest
selene~=2.0.0rc4
selenium~=4.10.0
allure-pytest